% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bib.R
\name{bib_to_list}
\alias{bib_to_list}
\title{Convert bib file to formatted list for Markdown output}
\usage{
bib_to_list(bib, arrange_by = "category")
}
\arguments{
\item{bib}{\verb{[character(1)]} Path to bibtex \code{.bib} file.}

\item{arrange_by}{\verb{[character()]} All lowercase biblatex field names to sort by. Passed to \code{dplyr::arrange()}.}
}
\value{
\code{character} vector with one element per entry in \code{bib}.
}
\description{
Convert bib file to formatted list for Markdown output
}
