% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-slides.R
\name{check_all_slides}
\alias{check_all_slides}
\title{Compile and compare all the slides}
\usage{
check_all_slides(
  lectures_tbl = collect_lectures(filter_lectures = lectures()),
  pre_clean = FALSE,
  create_comparison_pdf = TRUE,
  parallel = TRUE,
  thresh_psnr = 40,
  dpi_check = 50,
  dpi_out = 100,
  pixel_tol = 20,
  overwrite = FALSE
)
}
\arguments{
\item{lectures_tbl}{Must contain \code{tex} column. Defaults to \code{collect_lectures()}.}

\item{pre_clean}{\verb{[FALSE]}: Passed to \verb{[compile_slide()]}.}

\item{create_comparison_pdf, thresh_psnr, dpi_check, dpi_out, pixel_tol, overwrite}{Passed to \verb{[compare_slide()]}.}

\item{parallel}{\verb{[TRUE]} Whether to parallelize.
Uses \code{future.apply::future_lapply} with \code{future::plan("multisession")}.}
}
\value{
Invisibly: An expanded \code{lectures_tbl} with check results
Also saves output at \code{slide_check_cache.rds}.
}
\description{
Compile and compare all the slides
}
