% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_docker}
\alias{check_docker}
\title{Check if docker can be used}
\usage{
check_docker(strictness = c("warning", "error", "none"))
}
\arguments{
\item{strictness}{\verb{["warning"]} Wether to emit a warning, \code{"error"}, or nothing (\code{"none"}) if docker can not be used.}
}
\value{
Invisibly: \code{TRUE} if the docker seems to be running, \code{FALSE} otherwise, and an error if \code{strict} and the tool is not found.
}
\description{
\itemize{
\item \code{docker} needs to be in \verb{$PATH}
\item \code{docker} daemon (or compatible runtime) needs to be running
}
}
\examples{
check_docker(strictness = "none")
}
