% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-slides.R
\name{check_slides_many}
\alias{check_slides_many}
\title{Compile and compare many slides}
\usage{
check_slides_many(
  lectures_tbl = collect_lectures(),
  pre_clean = FALSE,
  compare_slides = FALSE,
  create_comparison_pdf = FALSE,
  parallel = TRUE,
  thresh_psnr = 40,
  dpi_check = 50,
  dpi_out = 100,
  pixel_tol = 20,
  overwrite = FALSE
)
}
\arguments{
\item{lectures_tbl}{Must contain \code{tex} column. Defaults to \code{\link[=collect_lectures]{collect_lectures()}}.}

\item{pre_clean}{\verb{[FALSE]}: Passed to \code{\link[=compile_slide]{compile_slide()}}.}

\item{compare_slides}{\verb{[FALSE]} If \code{TRUE}, run \code{\link[=compare_slide]{compare_slide()}} on the slide iff the compile check passed.}

\item{create_comparison_pdf, thresh_psnr, dpi_check, dpi_out, pixel_tol, overwrite}{Passed to \code{\link[=compare_slide]{compare_slide()}}.}

\item{parallel}{\verb{[TRUE]} Whether to parallelize.
Uses \link[future.apply:future_lapply]{future.apply::future_lapply} with \code{future::plan("multisession")}.}
}
\value{
Invisibly: An expanded \code{lectures_tbl} with check results
Also saves output at \code{slide_check_cache.rds}.
}
\description{
Wrapper for \code{\link[=check_slides_single]{check_slides_single()}}.
}
