% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-slides.R
\name{check_slides_single}
\alias{check_slides_single}
\title{Compile and compare a slide chunk}
\usage{
check_slides_single(
  slide_file,
  pre_clean = FALSE,
  compare_slides = FALSE,
  create_comparison_pdf = FALSE,
  thresh_psnr = 40,
  dpi_check = 50,
  dpi_out = 100,
  pixel_tol = 20,
  overwrite = FALSE
)
}
\arguments{
\item{slide_file}{\verb{[character(1)]} Name of a (single) slide, with or without \code{.tex} extension. See examples of \code{\link[=find_slide_tex]{find_slide_tex()}}.
Can also be a direct file path to enable use of this function outside rigid folder hierarchy.}

\item{pre_clean}{\verb{[FALSE]}: Passed to \code{\link[=compile_slide]{compile_slide()}}.}

\item{compare_slides}{\verb{[FALSE]} If \code{TRUE}, run \code{\link[=compare_slide]{compare_slide()}} on the slide iff the compile check passed.}

\item{create_comparison_pdf, thresh_psnr, dpi_check, dpi_out, pixel_tol, overwrite}{Passed to \code{\link[=compare_slide]{compare_slide()}}.}
}
\value{
A \code{data.frame} with columns
\itemize{
\item \code{tex}: Same as \code{slide_file} argument.
\item \code{compile_check}: \code{logical} indicating whether \code{\link[=compile_slide]{compile_slide()}} passed
\item \code{compare_check}: \code{logical} indicating whether \code{\link[=compare_slide]{compare_slide()}} passed ( \code{NA} if \code{compare_slides = FALSE} or \code{\link[=compile_slide]{compile_slide()}} did not pass)
\item \code{compare_check_note}: Note from \code{\link[=compare_slide]{compare_slide()}} indicating number and nature of differences.
\item \code{compare_check_raw}: More verbose form of \code{compare_check_note}.
\item \code{compile_note}: If there are compilation errors, the error messages from \code{\link[=compile_slide]{compile_slide()}} are included (see also \code{\link[=check_log]{check_log()}}).
}
}
\description{
Use \code{\link[=check_slides_many]{check_slides_many()}} to check many slides.
}
\examples{
\dontrun{
check_slides_single(slide_file = "slides-basics-whatisml", pre_clean = TRUE, compare_slides = TRUE)
}
}
