% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_system_tool}
\alias{check_system_tool}
\title{Simple check for availability of system tools}
\usage{
check_system_tool(x, strictness = c("warning", "error", "none"))
}
\arguments{
\item{x}{Name of a binary, e.g. \code{convert} for ImageMagick or \code{brew} for Homebrew on macOS.}

\item{strictness}{\verb{["warning"]} Wether to emit a warning, \code{"error"}, or nothing (\code{"none"}) if the tool is not found.}
}
\value{
\code{TRUE} if the tool is find, \code{FALSE} otherwise, and an error if \code{strict} and the tool is not found.
}
\description{
Can be used to verify if a tool (e.g. \code{convert}) is in \verb{$PATH} and findable from within R.
Sometimes a tool is in \verb{$PATH} in regular shell sessions but not within R.
}
\examples{
check_system_tool("diff-pdf", strictness = "none")
}
