% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bib.R
\name{clean_tex_markup}
\alias{clean_tex_markup}
\title{Clean up tex markup from biblatex entries}
\usage{
clean_tex_markup(x)
}
\description{
\itemize{
\item Removes \\ used for escaping
\item Substitutes \verb{\\emph\{foobar\}} with \verb{*foobar*}
\item Replaces \code{{foo}} with foo
}
}
\keyword{internal}
