% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectures-slides.R
\name{collect_lectures}
\alias{collect_lectures}
\title{Assemble table of lecture slides}
\usage{
collect_lectures(
  lectures_path = here::here(),
  filter_lectures = NULL,
  exclude_slide_subdirs = c("attic", "rsrc", "all", "figure_man", "figures_tikz",
    "figure", "tex", "backup"),
  exclude_slide_names = "chapter-order"
)
}
\arguments{
\item{lectures_path}{Path containing lecture_* directories. Defaulting to \code{here::here()}.}

\item{filter_lectures}{\verb{[NULL]}: Vector of lecture repo names to filter table by, e.g. \code{"lecture_i2ml"}.
Defaults to reading \code{LECTURES_INCLUDE} file via \code{\link[=lectures]{lectures()}}.}

\item{exclude_slide_subdirs}{Exclude slides/ subfolders, e.g. \code{c("attic", "rsrc", "all")}.}

\item{exclude_slide_names}{Exclude slides matching these names exactly, e.g. \code{"chapter-order"} (default).}
}
\value{
A data.frame with one row per slide .tex file.
}
\description{
Assemble table of lecture slides
}
\examples{
collect_lectures()
}
