% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-slides.R
\name{compare_slide}
\alias{compare_slide}
\title{Compare slide PDF against reference PDF}
\usage{
compare_slide(
  slide_file,
  verbose = TRUE,
  create_comparison_pdf = FALSE,
  thresh_psnr = 20,
  dpi_check = 50,
  eps_signif = 0.5,
  dpi_out = 100,
  pixel_tol = 50,
  view = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{slide_file}{\verb{[character(1)]} Name of a (single) slide, with or without \code{.tex} extension. See examples of \code{\link[=find_slide_tex]{find_slide_tex()}}.
Can also be a direct file path to enable use of this function outside rigid folder hierarchy.}

\item{verbose}{\verb{[TRUE]}: Print additional output to the console.}

\item{create_comparison_pdf}{\verb{[FALSE]}: Use \code{diff-pdf} to create a comparison PDF at \verb{./comparison/<slide-name>.pdf}.
The PDF contains only the slide pages with detected differences, highlighted in red/blue.}

\item{thresh_psnr}{\verb{[20]}: PSNR threshold for difference detection in \code{diff-pdf-visually}.
Higher is more sensitive.}

\item{dpi_check}{\verb{[50]} Resolution for rasterised files used by both \code{diff-pdf-visually}.
Lower is more coarse.}

\item{eps_signif}{\verb{[0.5]} Significance threshold for difference detection in \code{diff-pdf-visually} to force fewer false-positives.}

\item{dpi_out}{\verb{[100]} Resolution for output PDF produced by \code{diff-pdf}.
Lower values will lead to very pixelated diff PDFs.}

\item{pixel_tol}{\verb{[50]} Per-page pixel tolerance for comparison used by \code{diff-pdf}.}

\item{view}{\verb{[FALSE]} For interactive use: Opens window showing comparison diff.}

\item{overwrite}{\verb{[FALSE]} Re-creates output diff PDF even if it already exists and appears up to date.}
}
\value{
Invisibly: A list of results:
\itemize{
\item \code{passed}: TRUE indicates a successful comparison, FALSE a failure.
\item \code{reason}: Shorthand reason for a failing comparison.
\item \code{pages}: Vector of pages with differences.
\item \code{output}: Full output produced by diff-pdf-visually.
}

Also prints a summary, e.g.:

"! slides-cart-computationalaspects: Changes detected in pages: 11 and 12 (signif.: 0 and 0)"
}
\description{
A compiled .tex file such as \code{slides/gaussian-processes/slides-gp-bayes-lm.pdf} will be compared with
an assumed to be accepted / known good version at \code{slides-pdf/slides-gp-bayes-lm.pdf} if it exists.
}
\details{
First uses \code{diff-pdf-visually} to check slides for differences based on the PSNR of rasterised versions
of the slides (adjustable with \code{thresh_psnr}), and then uses \code{diff-pdf} to create a PDF highlighting
the differences if \code{create_comparison_pdf} is \code{TRUE}.

This only re-runs \code{diff-pdf} to create a comparison PDF file if the
output file under \verb{./comparison/<slide-name>.pdf} is more recent than both of the input PDF files.
That way you can safely re-run this function repeatedly without worrying about computational overhead.
}
\note{
Uses \code{diff-pdf-visually} and \code{diff-pdf} under the hood, you may need to adjust your $PATH.
}
\examples{
\dontrun{
# The "normal" way: A .tex file name
compare_slide("slides-cart-computationalaspects.tex")

# Also acceptable: A full path (absolute or relative), convenient for scripts
compare_slide("lecture_advml/slides/gaussian-processes/slides-gp-bayes-lm.tex")

# Lazy way: No extension, just a name
compare_slide("slides-cart-predictions")

# Whoopsie supplied name of PDF instead no biggie I got u
compare_slide("slides-forests-proximities.pdf")

compare_slide("slides-boosting-cwb-basics")
}
}
