% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_dir}
\alias{ensure_dir}
\title{Ensure a directory exists}
\usage{
ensure_dir(x)
}
\arguments{
\item{x}{(\code{character(1)}) A file or directory path.}
}
\value{
\code{TRUE} if the file exists, \code{FALSE} otherwise
}
\description{
Creates the directory \code{x} or the directory containing the file \code{x}.
}
