% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bib.R
\name{format_authors}
\alias{format_authors}
\title{Condense list of authors from bilatex entry}
\usage{
format_authors(author_list)
}
\description{
\itemize{
\item NAs are replaced by ---
\item Two authors are separated by &
\item More than two authors are replaced by First Author et al.
}
}
\keyword{internal}
