% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_lecheck}
\alias{install_lecheck}
\title{Install the lecheck cli tool}
\usage{
install_lecheck(path = "~/.local/bin", overwrite = TRUE)
}
\arguments{
\item{path}{\verb{["~/.local/bin"]} Path to symlink the tool to. Must exist and be writable.}

\item{overwrite}{\verb{[TRUE]} Overwrite any existing symlink.}
}
\value{
\code{FALSE} if a symlink already exists and is not overwritten.
Otherwise: The path to the symlink is returned.
}
\description{
\code{path} should be in the user's \verb{$PATH} to make the tool usable in shell sessions.
}
\examples{
if (FALSE) {
install_lecheck()

# Would only work if the R session is started under a user that can write to /usr/local/bin
install_lecheck("/usr/local/bin")
}
}
