% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexmk.R
\name{latexmk_docker}
\alias{latexmk_docker}
\title{Run dockerized latexmk}
\usage{
latexmk_docker(slide_file, verbose = TRUE, tag = "TL2023-historic")
}
\arguments{
\item{slide_file}{\verb{[character(1)]} Name of a (single) slide, with or without \code{.tex} extension. See examples of \code{\link[=find_slide_tex]{find_slide_tex()}}.}

\item{verbose}{\verb{[TRUE]}: Print output from \code{docker}/\code{latexmk} to console.}
}
\description{
This uses the docker image from \url{https://gitlab.com/islandoftex/images/texlive}.
The defautl uses tag \code{TL2023-historic} for TeXLive 2023.
}
\details{
You will need to install docker or podman or some other compatible runtime on your system beforehand.

The docker command run by this is equivalent to something like

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{cd path/to/lecture_i2ml/slides/ml-basics

CWD=$(basename $\{PWD\})
LECTURE=$(dirname $(dirname $\{PWD\}))

docker run -i --rm --user $(id -u) --name latex \\
  -v "$\{LECTURE\}":/usr/src/app:z \\
  -w "/usr/src/app/slides/$\{CWD\}" \\
  registry.gitlab.com/islandoftex/images/texlive:TL2023-historic \\
  latexmk -pdf -halt-on-error slides-basics-data.tex
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
latexmk_docker("slides-advriskmin-bias-variance-decomposition.tex")
}

}
