% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexmk.R
\name{latexmk_system}
\alias{latexmk_system}
\title{Run latexmk}
\usage{
latexmk_system(
  slide_file,
  verbose = TRUE,
  log_stdout = "",
  log_stderr = "",
  supervise = TRUE
)
}
\arguments{
\item{slide_file}{\verb{[character(1)]} Name of a (single) slide, with or without \code{.tex} extension. See examples of \code{\link[=find_slide_tex]{find_slide_tex()}}.
Can also be a direct file path to enable use of this function outside rigid folder hierarchy.}

\item{verbose}{\verb{[TRUE]}: Print output from \code{docker}/\code{latexmk} to console.}

\item{log_stdout, log_stderr}{\verb{[""]}: Path to write stdout/stderr log to.
Discarded if \code{NULL} or inherited from main R process if \code{""}.
\code{stderr} can be redirected to \code{stdout} with \code{"2>&1"}.}

\item{supervise}{\verb{[TRUE]}: Passed to \code{\link[processx:process]{processx::process()}}'s \verb{$new()}.}
}
\value{
A \code{\link[processx:process]{processx::process()}} object.
}
\description{
\code{latexmk} needs to be in \verb{$PATH} for this to work.
}
\note{
This utility is usually invoked by \code{\link[=compile_slide]{compile_slide()}}.
}
\examples{
\dontrun{
latexmk_system("slides-advriskmin-bias-variance-decomposition.tex")
}
}
