% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexmk.R
\name{latexmk_tinytex}
\alias{latexmk_tinytex}
\title{Run TinyTex's latexmk}
\usage{
latexmk_tinytex(slide_file, ...)
}
\arguments{
\item{slide_file}{\verb{[character(1)]} Name of a (single) slide, with or without \code{.tex} extension. See examples of \code{\link[=find_slide_tex]{find_slide_tex()}}.}

\item{...}{Arguments passed to \code{\link[tinytex:latexmk]{tinytex::latexmk()}}.}
}
\value{
\code{TRUE} if the output PDF exists, \code{FALSE} otherwise.
}
\description{
A thin wrapper around \code{\link[tinytex:latexmk]{tinytex::latexmk()}}.
}
\details{
TinyTex's \code{\link[tinytex:latexmk]{tinytex::latexmk()}} automatically installs missing LaTeX packages,
making it very useful.
This is just a thin wrapper run the command with
a changed working directory, as relative paths used in \code{preamble.tex} etc. require.
}
\note{
This utility is usually invoked by \code{\link[=compile_slide]{compile_slide()}}.
}
\examples{
\dontrun{
latexmk_tinytex("slides-advriskmin-bias-variance-decomposition.tex")
}
}
