% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectures-slides.R
\name{lectures}
\alias{lectures}
\title{Get included lectures}
\usage{
lectures()
}
\value{
A character vector, e.g. \code{c("lecture_i2ml", "lecture_advml")}, depending on \code{./include_lectures}
}
\description{
Uses central file \code{./include_lectures}, ignoring commented out lines.
Can be overridden with environment variable \verb{$include_lectures}.
If neither the environment variable nor the file exists, it defaults to listing
all lectures.
}
\examples{
\dontrun{
lectures()
}
}
