% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lese-package.R
\docType{package}
\name{lese-package}
\alias{lese}
\alias{lese-package}
\title{lese: SLDS Lecture Service Companion}
\description{
Provides infrastructure around SLDS lectures for maintenance and development.
}
\author{
\strong{Maintainer}: Lukas Burk \email{github@quantenbrot.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})

}
\keyword{internal}
