% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\alias{\%|0|\%}
\title{Default value for \code{NULL}}
\usage{
x \%||\% y

x \%|0|\% y
}
\arguments{
\item{x, y}{If \code{x} is NULL or length 0, will return \code{y}; otherwise returns \code{x}.}
}
\description{
This infix function makes it easy to replace \code{NULL}s with a default
value. It's inspired by the way that Ruby's or operation (\code{||})
works.
}
\examples{
1 \%||\% 2
NULL \%||\% 2
character(0) \%|0|\% ""
list() \%|0|\% ""
}
