% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_margin_token_file}
\alias{set_margin_token_file}
\title{Auto-set the with/without margin dummy file}
\usage{
set_margin_token_file(wd, margin = TRUE, token_name = "nospeakermargin.tex")
}
\arguments{
\item{wd}{Working directory (relative or absolute) where the file needs to be created.
This is the directory were the \code{.tex} file to be compiles is located.}

\item{margin}{\verb{[TRUE]} Whether to enable (\code{TRUE}, default) or disable (\code{FALSE}) the margin.}

\item{token_name}{\code{"nospeakermargin.tex"} If the name changes or needs to be flexible for testing
it can be adjusted, but typically the default is set in stone.}
}
\value{
Nothing
}
\description{
See \code{style/lmu-lecture.sty} where depending on the presence of an empty \code{.tex} file with a
specific name certain layout options are set to compile slides either in 16:9 with margins or
in 4:3 without a margin for the speaker.
}
\examples{
wd <- tempdir()

set_margin_token_file(wd = wd, margin = FALSE)
stopifnot("file exists when no margin set" = file.exists(file.path(wd, "nospeakermargin.tex")))

set_margin_token_file(wd = wd, margin = TRUE)
stopifnot("file is removed when margin set" = !file.exists(file.path(wd, "nospeakermargin.tex")))
}
