% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide-cache.R
\name{slide_cache_clean}
\alias{slide_cache_clean}
\title{Delete the slide check cache}
\usage{
slide_cache_clean()
}
\value{
\code{TRUE} (invisibly) if the file was deleted, \code{FALSE} if it
did not exist.
}
\description{
Removes the cache file created by \code{\link[=check_slides_many]{check_slides_many()}}, if it exists.
}
\examples{
\dontrun{
slide_cache_clean()
}
}
