% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{this_repo_status}
\alias{this_repo_status}
\title{Status of the service repo checkout
Same as \code{lecture_status_local()} but for this service repo}
\usage{
this_repo_status()
}
\value{
A \code{data.frame} similar to \code{lecture_status_local()}.
}
\description{
Status of the service repo checkout
Same as \code{lecture_status_local()} but for this service repo
}
\keyword{internal}
