% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-log.R
\name{check_log}
\alias{check_log}
\title{Check latexmk logs for common errors}
\usage{
check_log(slide_file, before = 0, after = 1)
}
\arguments{
\item{slide_file}{Name of a slide file, e.g. \code{"slides-gp-basic-3"}, with optional file extension.}

\item{before, after}{\verb{[integer(1)]} Number of log lines to display before and after the line found via regex.
Defaults to 0 lines before, 1 line after.}
}
\description{
Check latexmk logs for common errors
}
\examples{
\dontrun{
check_log("slides-gp-basic-3")
# "976: ! Undefined control sequence.\n976: l.9 \\newcommandiscrete\n"
}
}
