% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_slide.R
\name{clean_slide}
\alias{clean_slide}
\title{Clean output for a single .tex file}
\usage{
clean_slide(slide_file, keep_pdf = FALSE, verbose = FALSE, check_status = TRUE)
}
\arguments{
\item{slide_file}{\verb{[character(1)]} Name of a (single) slide, with or without \code{.tex} extension. See examples of \code{\link[=find_slide_tex]{find_slide_tex()}}.
Can also be a direct file path to enable use of this function outside rigid folder hierarchy.}

\item{keep_pdf}{\verb{[FALSE]}: Keep the PDF file.}

\item{verbose}{\verb{[TRUE]}: Print additional output to the console.}

\item{check_status}{\verb{[TRUE]}: Wait for \code{latexmk} to finish and return the exit status. Not supported for \code{method = "tinytex"}.}
}
\value{
Invisibly:
\itemize{
\item If \code{check_status}, \code{TRUE} if the exit code is 0, \code{FALSE} otherwise.
\item If \code{check_status} is \code{FALSE}, \code{NULL} is returned.
}
}
\description{
Uses \verb{latexmk -C <slide_file>}, also removing the PDF file.
Uses \verb{latexmk -c <slide_file>} to keep the PDF file.
}
\examples{
\dontrun{
# Create the PDF
compile_slide("slides-cart-computationalaspects.tex")

# Remove the PDF and other output
clean_slide("slides-cart-computationalaspects.tex")
}
}
