% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-slides.R
\name{compile_slide}
\alias{compile_slide}
\title{Compile a single .tex file}
\usage{
compile_slide(
  slide_file,
  pre_clean = TRUE,
  margin = TRUE,
  check_status = TRUE,
  verbose = TRUE,
  log = FALSE
)
}
\arguments{
\item{slide_file}{\code{character(1)}: A single slide .tex file (see examples).}

\item{pre_clean}{\verb{[TRUE]}: Run \code{clean_slide()} beforehand, ensuring a clean slate.}

\item{margin}{\verb{[TRUE]} By default renders slides with margin. Otherwise a 4:3 slide is
rendered.}

\item{check_status}{\verb{[TRUE]}: Wait for \code{make} to finish and return the exit status.}

\item{verbose}{\verb{[TRUE]}: Print additional output to the console.}

\item{log}{\verb{[FALSE]}: Write stdout and stderr logs to \verb{./logs/}.}
}
\value{
Invisibly: A list with entries
\itemize{
\item passed: TRUE indicates a successful compilation, FALSE a failure.
\item log: Absolute path to the log file in case of a non-zero exit status.
}
}
\description{
Compile a single .tex file
}
\examples{
\dontrun{
# The "normal" way: A .tex file name
compile_slide("slides-cart-computationalaspects.tex")

# Also acceptable: A full path (absolute or relative), convenient for scripts
compile_slide("lecture_advml/slides/gaussian-processes/slides-gp-bayes-lm.tex")

# Lazy way: No extension, just a name
compile_slide("slides-cart-predictions")
}
}
