% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectures-slides.R
\name{find_slide_tex}
\alias{find_slide_tex}
\title{Find a slide set across all lectures}
\usage{
find_slide_tex(slide_file, lectures_tbl = NULL)
}
\arguments{
\item{slide_file}{\verb{[character(1)]} Name of a (single) slide, with or without \code{.tex} extension. See examples of \code{\link[=find_slide_tex]{find_slide_tex()}}.
Can also be a direct file path to enable use of this function outside rigid folder hierarchy.}

\item{lectures_tbl}{\verb{[NULL]} Must contain \code{tex} column. Will use \code{collect_lectures()} if not set and \code{slide_file} is not an existing file path.}
}
\description{
Lectures need to be stored locally in the current directory with regular names like  \code{lecture_i2ml}.
It is strongly assumed that slide names such as \code{slides-cart-predictions.tex} are unique across all lectures.
}
\examples{
\dontshow{if (fs::dir_exists(here::here("lecture_i2ml"))) withAutoprint(\{ # examplesIf}
# The "normal" way: A .tex file name
str(find_slide_tex(slide_file = "slides-cart-computationalaspects.tex"))

# Also acceptable: A full path (absolute or relative), convenient for scripts
str(find_slide_tex(slide_file = "lecture_i2ml/slides/cart/slides-cart-predictions.tex"))

# Lazy way: No extension, just a name
str(find_slide_tex(slide_file = "slides-cart-predictions"))

# Can also get the .tex file for a .pdf
str(find_slide_tex(slide_file = "slides-cart-predictions.pdf"))
\dontshow{\}) # examplesIf}
}
