% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectures-slides.R
\name{find_slide_tex}
\alias{find_slide_tex}
\title{Find a slide set across all lectures}
\usage{
find_slide_tex(lectures_tbl = collect_lectures(), slide_file)
}
\arguments{
\item{lectures_tbl}{Must contain \code{tex} column. Defaults to \code{collect_lectures()}.}

\item{slide_file}{Name of a (single) slide, with or without \code{.tex} extension. See examples.}
}
\description{
Lectures need to be stored locally in the current directory with regular names like  \code{lecture_i2ml}.
It is strongly assumed that slide names such as \code{slides-cart-predictions.tex} are unique across all lectures.
}
\examples{
\dontrun{
# The "normal" way: A .tex file name
find_slide_tex("slides-cart-computationalaspects.tex")

# Also acceptable: A full path (absolute or relative), convenient for scripts
find_slide_tex("lecture_advml/slides/gaussian-processes/slides-gp-bayes-lm.tex")

# Lazy way: No extension, just a name
find_slide_tex("slides-cart-predictions")
}
}
