% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectures-slides.R
\name{lectures}
\alias{lectures}
\title{Read included lectures from one central file, ignoring commented out lines}
\usage{
lectures()
}
\value{
A character vector, e.g. \code{c("lecture_i2ml", "lecture_advml")}, depending on \code{./include_lectures}
}
\description{
Can be overridden with environment variable \verb{$include_lectures}.
}
\examples{
\dontrun{
lectures()
}
}
