% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_status.R
\name{this_repo_status}
\alias{this_repo_status}
\title{Service repo checkout status}
\usage{
this_repo_status()
}
\description{
Same as \code{repo_status()} but for this service repo
}
\examples{
if (FALSE) this_repo_status()
}
